+++ qb_device_driver/CHANGELOG.rst
@@ -2,6 +2,9 @@
 Changelog for package qb_device_driver
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+1.1.0 (2017-11-24)
+------------------
+
 1.0.8 (2017-06-27)
 ------------------
 * Fix C++11 support for cmake version less than 3.1